feat(crates/compiler): Preliminary
