+++ crates/compiler/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "compiler"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+typst = "0.11.1"
+typst-svg = "0.11.1"
+ttf-parser = "0.21.1"
+comemo = "0.4.0"
+typst-assets = {version = "0.11.1", features = ["fonts"]}
